Move pkg/installer to luet-build
